ReadaheadRandomAccessFile -- userspace readahead
Summary:
ReadaheadRandomAccessFile acts as a transparent layer on top of RandomAccessFile. When a Read() request is issued, it issues a much bigger request to the OS and caches the result. When a new request comes in and we already have the data cached, it doesn't have to issue any requests to the OS.

We add ReadaheadRandomAccessFile layer only when file is read during compactions.

Test Plan: make check

Reviewers: MarkCallaghan, sdong

Subscribers: dhruba, leveldb

Differential Revision: https://reviews.facebook.net/D45105